use platform definition for alternate config of Knoscillator, rather than templated patch class.
