+++ Source/KnoscillatorPatch.hpp
@@ -30,49 +30,20 @@ DESCRIPTION:
 */
 #pragma once
 
-#include "Patch.h"
+#include "MonochromeScreenPatch.h"
 #include "MidiMessage.h"
 #include "VoltsPerOctave.h"
 #include "SmoothValue.h"
 #include "vessicle/Knoscillator.h"
 #include "vessicle/Projector.h"
+#include "KnoscillatorParamIds.hpp"
+#include "CircularBuffer.h"
 
-struct KnoscillatorParameterIds
+class KnoscillatorPatch : public MonochromeScreenPatch
 {
-  PatchParameterId inPitch;
-  PatchParameterId inMorph;
-  PatchParameterId inKnotP;
-  PatchParameterId inKnotQ;
-  PatchParameterId inKnotS;
-  PatchParameterId inDetuneP;
-  PatchParameterId inDetuneQ;
-  PatchParameterId inDetuneS;
-  PatchParameterId inRotateX;
-  PatchParameterId inRotateY;
-  PatchParameterId inRotateZ;
-  PatchParameterId inRotateXRate;
-  PatchParameterId inRotateYRate;
-  PatchParameterId inRotateZRate;
-  PatchParameterId inNoiseAmp;
-  PatchParameterId inFMRatio;
-  PatchParameterId inZoom;
-
-  PatchParameterId outRotateX;
-  PatchParameterId outRotateY;
-  PatchParameterId outRotateZ;
-
-  PatchButtonId inFreezeP;
-  PatchButtonId inFreezeQ;
-  PatchButtonId outRotateXGate;
-  PatchButtonId outRotateYGate;
-  PatchButtonId outRotateZGate;
-};
-
-template<typename T, typename PatchClass = Patch>
-class KnoscillatorPatch : public PatchClass
-{
-  using KnoscilGen = Knoscillator<T, true>;
-  using Camera = Projector<T>;
+  using KnoscilGen = Knoscillator<float, true>;
+  using Camera = Projector<float>;
+  
 protected:
   KnoscillatorParameterIds params;
   VoltsPerOctave hz;
@@ -95,25 +66,32 @@ private:
   float rotateOffY;
   float rotateOffZ;
   float rotateOffSmooth;
+  
+  // for Genius
+  size_t drawCount;
+  float  knotPhase;
+  CircularFloatBuffer* scopeLeft;
+  CircularFloatBuffer* scopeRight;
+  MonochromeScreenBuffer backBuffer;
 
 public:
-  using PatchClass::registerParameter;
-  using PatchClass::getParameterValue;
-  using PatchClass::setParameterValue;
-  using PatchClass::getSampleRate;
-  using PatchClass::isButtonPressed;
-  using PatchClass::setButton;
-  using PatchClass::getBlockSize;
-
-  explicit KnoscillatorPatch(const KnoscillatorParameterIds& paramIds)
-    : PatchClass()
-    , params(paramIds), hz(true), midinote(0), tune(0.5f, -6.0f), knotP(0.9f, 2), knotQ(0.9f, 1)
+  explicit KnoscillatorPatch()
+    : params(knoscillatorParamIds), hz(true), midinote(0), tune(0.5f, -6.0f)
+    , knotP(0.9f, 2), knotQ(0.9f, 1)
     , morph(0.9f, 0), fmRatio(0.9f, 2.0f)
     , rotateOffX(0), rotateOffY(0), rotateOffZ(0)
     , rotateOffSmooth(4.0f * vessl::math::pi<float>() * 2 / getSampleRate())
+    , drawCount(0), knotPhase(0), backBuffer(128,64)
   {
     knoscil = KnoscilGen::create(getSampleRate());
     camera = Camera::create();
+    
+#ifdef OWL_GENIUS
+    const size_t bufferSize = getBlockSize() * 128;
+    scopeLeft = CircularFloatBuffer::create(bufferSize);
+    scopeRight = CircularFloatBuffer::create(bufferSize);
+    backBuffer.setBuffer(new uint8_t[backBuffer.getWidth()*backBuffer.getHeight()/8]);
+#endif
 
     registerParameter(params.inPitch, "Pitch");
     registerParameter(params.inMorph, "Morph");
@@ -174,10 +152,22 @@ public:
     setParameterValue(params.inZoom, 1);
   }
 
-  virtual ~KnoscillatorPatch()
+  virtual ~KnoscillatorPatch() override
   {
     KnoscilGen::destroy(knoscil);
     Camera::destroy(camera);
+    
+#ifdef OWL_GENIUS
+    CircularFloatBuffer::destroy(scopeLeft);
+    CircularFloatBuffer::destroy(scopeRight);
+    delete[] backBuffer.getBuffer();
+#endif
+  }
+  
+  // returns CPU% as [0,1] value
+  float getElapsedTime()
+  {
+    return getElapsedCycles() / getBlockSize() / 10000.0f;
   }
 
   void processMidi(MidiMessage msg)
@@ -190,6 +180,24 @@ public:
 
   void processAudio(AudioBuffer& audio) override
   {
+#ifdef OWL_GENIUS
+    {
+      const int size = audio.getSize();
+      FloatArray inLeft = audio.getSamples(0);
+      const float phaseStep = 1.0f / getSampleRate();
+      for (int i = 0; i < size; ++i)
+      {
+        const float freq = hz.getFrequency(inLeft[i]);
+        knotPhase += freq * phaseStep;
+        if (knotPhase >= 1)
+        {
+          knotPhase -= 1;
+          drawCount = scopeLeft->getReadCapacity() + i;
+        }
+      }
+    }
+#endif
+    
     FloatArray left = audio.getSamples(LEFT_CHANNEL);
     FloatArray right = audio.getSamples(RIGHT_CHANNEL);
 
@@ -302,5 +310,58 @@ public:
       setButton(params.outRotateYGate, gateY);
       setButton(params.outRotateZGate, gateZ);
     }
+    
+#ifdef OWL_GENIUS
+    {
+      scopeLeft->write(audio.getSamples(0), audio.getSize());
+      scopeRight->write(audio.getSamples(1), audio.getSize());
+    }
+#endif
+  }
+  
+  void processScreen(MonochromeScreenBuffer& screen) override
+  {
+#ifdef OWL_GENIUS
+    const int displayHeight = screen.getHeight() - 18;
+    const int cy = displayHeight / 2;
+    const int cx = screen.getWidth() / 2;
+    const int sz = screen.getHeight() / 2;
+
+    if (drawCount)
+    {
+      backBuffer.clear();
+      float t1 = getElapsedTime();
+      size_t count = vessl::math::min(drawCount, scopeLeft->getSize() / 2);
+      drawCount -= count;
+      while (count--)
+      {
+        int x = cx + scopeLeft->read()*sz;
+        int y = cy + scopeRight->read()*sz;
+        backBuffer.setPixel(x, y, WHITE);
+      }
+      float t2 = getElapsedTime();
+
+      //backBuffer.setCursor(0, 8);
+      //backBuffer.print(t2 - t1);
+    }
+    // copy pixels one at a time so we don't overwrite
+    // the selected parameters displayed at the bottom of the screen
+    for (int x = 0; x < screen.getWidth(); ++x)
+    {
+      for (int y = 0; y < displayHeight; ++y)
+      {
+        auto c = screen.getPixel(x, y) | backBuffer.getPixel(x, y);
+        screen.setPixel(x, y, c);
+      }
+    }
+    
+    // screen.setCursor(0, 10);
+    // screen.print("dt "); screen.print((int)knoscil->knot().pi()); 
+    // screen.print(" sr " ); screen.print(knoscil->knot().sr()); screen.print("\n");
+    // screen.print("pp "); screen.print((int)knoscil->knot().pp()); screen.print("\n");
+    // screen.print("pq "); screen.print((int)knoscil->knot().pq()); screen.print("\n");
+    // screen.print("pz "); screen.print((int)knoscil->knot().pz()); screen.print("\n");
+    // screen.print(knoscil->knot().frequency().readAnalog());
+#endif
   }
 };
